ARCHITECTURAL/ArchiGPT_0.2/datasets/initial/10_RentYourExpert/rentyourexpert-main/backend/cust_login_ms/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
mysql-connector-python
Flask-JWT-Extended==4.2.3
bcrypt==3.2.0
Werkzeug==2.2.2